requirement.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
Deprecated==1.2.14
Flask==2.3.3
Flask-Limiter==3.5.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
google-images-download==2.8.0
google-search-results==2.4.2
google-serp-api==1.0.4
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
limits==3.6.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.0
openai==0.28.1
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
Pillow==9.5.0
protobuf==4.24.3
pyarrow==13.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.16.1
PySocks==1.7.1
python-dateutil==2.8.2
python-pptx==0.6.22
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.5.3
rpds-py==0.10.3
selenium==4.13.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
streamlit==1.27.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
XlsxWriter==3.1.5
yarl==1.9.2
zipp==3.17.0
